Update m5 calculation.
Adds an update to adjust for case where sky noise trades off as dominant noise source with readnoise. This upgrades the accuracy of the m5 calculation to within a few percent of what would be calculated directly with full SNR calculation. This makes the biggest change in the u band, where the value of the sky brightness means that the observation is read-noise vs sky-noise dominated.